+++ src/lib/periph/uart/uart.h
@@ -19,6 +19,7 @@
 #pragma once
 
 #include <stdbool.h>
+#include <stddef.h>
 #include <stdint.h>
 
 #include "core/queue.h"
@@ -36,7 +37,7 @@ typedef void (*uart_rx_cb)(UartState_t *s, void *user_data, char c);
 struct UartState_t_s {
   bool initted;
   uint8_t uart_id;
-  uint16_t max_tx_len;
+  size_t max_tx_len;
 
   // send
   union {
@@ -63,7 +64,7 @@ void uart_init(UartState_t *u, uint8_t uart_id, uint32_t baud);
 void uart_start_rx(UartState_t *u, uart_rx_cb rx_cb, void *user_data);
 
 // Sends binary data to the UART, specified with a length.
-void uart_write(UartState_t *u, const void *buf, uint16_t len);
+void uart_write(UartState_t *u, const void *buf, size_t len);
 
 // Sends a null-terminated string to the UART.
 void uart_print(UartState_t *u, const char *s);
